refactor auth util functions in tests
