+++ test/tools/spec-runner/context.js
@@ -1,7 +1,8 @@
 'use strict';
 const { expect } = require('chai');
-const { resolveConnectionString, extractAuthString } = require('./utils');
+const { resolveConnectionString } = require('./utils');
 const { ns } = require('../../../src/utils');
+const { extractAuthFromConnectionString } = require('../utils');
 
 class Thread {
   constructor() {
@@ -81,7 +82,9 @@ class TestRunnerContext {
     if (config.topologyType === 'Sharded') {
       this.failPointClients = config.options.hostAddresses.map(proxy => {
         const authString =
-          process.env.AUTH === 'auth' ? `${extractAuthString(process.env.MONGODB_URI)}@` : '';
+          process.env.AUTH === 'auth'
+            ? `${extractAuthFromConnectionString(process.env.MONGODB_URI)}@`
+            : '';
         return config.newClient(
           `mongodb://${authString}${proxy.host}:${proxy.port}/${
             process.env.AUTH === 'auth' ? '?authSource=admin' : ''
