+++ test/tools/spec-runner/utils.js
@@ -1,27 +1,6 @@
 'use strict';
 
-function getIndicesOfAuthInUrl(connectionString) {
-  const doubleSlashIndex = connectionString.indexOf('//');
-  const atIndex = connectionString.indexOf('@');
-
-  if (doubleSlashIndex === -1 || atIndex === -1) {
-    return null;
-  }
-
-  return {
-    start: doubleSlashIndex + 2,
-    end: atIndex
-  };
-}
-
-function extractAuthString(connectionString) {
-  const indices = getIndicesOfAuthInUrl(connectionString);
-  if (!indices) {
-    return null;
-  }
-
-  return connectionString.slice(indices.start, indices.end);
-}
+const { extractAuthFromConnectionString } = require('../utils');
 
 function resolveConnectionString(configuration, spec, context) {
   const isShardedEnvironment = configuration.topologyType === 'Sharded';
@@ -30,7 +9,9 @@ function resolveConnectionString(configuration, spec, context) {
   const password = context && context.password;
   const authSource = (context && context.authSource) || 'admin';
   const authString =
-    process.env.AUTH === 'auth' ? `${extractAuthString(process.env.MONGODB_URI)}@` : '';
+    process.env.AUTH === 'auth'
+      ? `${extractAuthFromConnectionString(process.env.MONGODB_URI)}@`
+      : '';
   const connectionString =
     isShardedEnvironment && !useMultipleMongoses
       ? `mongodb://${authString}${configuration.host}:${configuration.port}/${configuration.db}?directConnection=false&authSource=${authSource}`
@@ -38,4 +19,4 @@ function resolveConnectionString(configuration, spec, context) {
   return connectionString;
 }
 
-module.exports = { resolveConnectionString, getIndicesOfAuthInUrl, extractAuthString };
+module.exports = { resolveConnectionString };
